Simplify WASM
WASM no longer attempts to detect new lvars automatically; it now requires the
client to ask it.  Checking is now more efficient (it just checks from the previous
point), and is invoked every ~30 seconds by the client. (Some aircraft add lvars
sometime after loading up, so simply monitoring AircraftLoaded + FlightLoaded
aren't enough.)

+++ build-and-publish.ps1
@@ -1,9 +1,13 @@
-Remove-Item -Path .\BravoLights\bin\Release -Recurse
+if (Test-Path -Path .\BravoLights\bin\Release) {
+	Remove-Item -Path .\BravoLights\bin\Release -Recurse
+}
 
 dotnet publish -p:PublishProfile=FolderProfile
 $path = Resolve-Path "BravoLights\bin\Release\net5.0-windows\publish\BetterBravoLights.exe"
 
 $o = [system.diagnostics.fileversioninfo]::GetVersionInfo($path)
 Write-Output "Better Bravo Lights $($o.ProductVersion)" | Out-File -Encoding UTF8 BravoLights\bin\Release\net5.0-windows\publish\VERSION.txt
+
+Start-Process -Wait -FilePath "$env:MSFS_SDK\Tools\bin\fspackagetool.exe" -ArgumentList "MSFSWASMProject\BetterBravoLightsLVars.xml -outputdir BravoLights\bin\Release\net5.0-windows\publish"
 
 Compress-Archive -Path BravoLights\bin\Release\net5.0-windows\publish\* -Force -DestinationPath BetterBravoLights.zip
